fix: make IamAssumeAuthenticator immutable after initial construction
